requirements.txt
pycryptodome
yt-dlp
pillow
python-slugify
pycryptodomex
dnspython
google-api-python-client
google-auth-oauthlib
oauth2client==3.0.0
lxml
psutil
motor
python-dotenv
python-magic
requests
pyrofork
hachoir
pytz
google-auth-httplib2
tgcrypto
cryptography
protobuf==3.20.1
pymediainfo
Requests
titlecase
urllib3
xmltodict
expiringdict
tenacity
unidecode
pyopenssl